adds JSON encoding and decoding + its tests
