+++ src/test/java/modules/JsonEncoderTest.java
@@ -1,6 +1,7 @@
 package modules;
 
 import model.Entity;
+import model.fixture.TestEntity;
 import modules.codec.JsonEncoder;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -12,7 +13,7 @@ public class JsonEncoderTest {
     @Test
     public void Test(){
         JsonEncoder encoder = new JsonEncoder();
-        TestEntity entity = new TestEntity(16482, "dummy string", 3.14159);
+        TestEntity entity = new TestEntity(16482, "dummy string", 3.14159, "TestEntity");
         Entity entityChanged = encoder.decode(encoder.encode(entity));
         Assertions.assertEquals(entity, entityChanged);
         System.out.println(IdentifierFixture.NewIdentifier().toString());
